# specs/002-admin-dashboard/contracts/api.yaml
openapi: 3.0.3
info:
  title: Admin Dashboard API
  description: API REST pour l'application d'administration de la plateforme dropshipping
  version: 1.0.0
servers:
  - url: http://localhost:8000/admin
    description: Serveur de développement
  - url: https://api.dropshipping.local/admin
    description: Serveur de production
...
paths:
...
  /creatives:
    get:
      tags:
        - creatives
      summary: Liste des créatives
      description: Retourne la liste des créatives générées
      security:
        - bearerAuth: []
      parameters:
        - name: product_id
          in: query
          schema:
            type: integer
          description: Filtrer par produit
        - name: status
          in: query
          schema:
            type: string
            enum: [queued, processing, completed, failed]
          description: Filtrer par statut
      responses:
        '200':
          description: Liste des créatives
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Creative'
                  total:
                    type: integer
        '401':
          description: Non authentifié
  /orders:
    get:
      tags:
        - orders
      summary: Liste des commandes
      description: Retourne la liste des commandes avec filtres
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, shipped, delivered]
          description: Filtrer par statut
        - name: date_from
          in: query
          schema:
            type: string
            format: date
          description: Date de début
        - name: date_to
          in: query
          schema:
            type: string
            format: date
          description: Date de fin
      responses:
        '200':
          description: Liste des commandes
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  total:
                    type: integer
        '401':
          description: Non authentifié
  /orders/{order_id}/status:
    patch:
      tags:
        - orders
      summary: Mettre à jour le statut d'une commande
      description: Met à jour le statut d'une commande
      security:
        - bearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [pending, confirmed, shipped, delivered]
      responses:
        '200':
          description: Statut mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Statut invalide ou erreur
        '401':
          description: Non authentifié
        '404':
          description: Commande non trouvée
...
  /actions/creatives/generate/{product_id}:
    post:
      tags:
        - actions
      summary: Générer les créatives pour un produit
      description: Déclenche manuellement la génération de créatives pour un produit
      security:
        - bearerAuth: []
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Génération de créatives démarrée
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                    description: Identifiant du job de génération
                  status:
                    type: string
                    enum: [queued, started]
        '401':
          description: Non authentifié
        '404':
          description: Produit non trouvé
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Creative:
      type: object
      properties:
        id:
          type: integer
        product_id:
          type: integer
        video_type:
          type: string
          enum: [ugc, problem_solution, short_hook, carousel]
        status:
          type: string
          enum: [queued, processing, completed, failed]
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
    Order:
      type: object
      properties:
        id:
          type: integer
        product_id:
          type: integer
        product_name:
          type: string
        full_name:
          type: string
        phone:
          type: string
        quantity:
          type: integer
        status:
          type: string
          enum: [pending, confirmed, shipped, delivered]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time